URL RPC Data Parser
This tool is not specific to aggregating RPC data, however it is specializes for it with just a few features. 
The parse Variables are in lines 163,164,166,168, and 173; 
the script will sift through any HTML syntax given that it is in vetween '<'  and '>', placing only the outputs in the list for aggregation;
That list can be filtered with line 178.
at the end of the script, there is a use case specific for loop function that clarifies correct data, you are free to do with it as you please.